+++ praxis/layers/quantization/tests/combined_qkv_projection_fq_test.py
@@ -480,7 +480,8 @@ class CombinedQKVProjectionFQTest(quantization_test_util.QuantizationTestCase):
         quantization_type=QuantizationType.FQ,
         mode=QuantizationMode.TRAINING,
         weight_params=quantization_hparams.WeightQuantizationParams(
-            use_symmetric=is_weight_symmetric
+            use_symmetric=is_weight_symmetric,
+            calculation_dtype=jnp.bfloat16,
         ),
     )
     p_f = pax_fiddle.Config(
@@ -538,7 +539,8 @@ class CombinedQKVProjectionFQTest(quantization_test_util.QuantizationTestCase):
         quantization_type=QuantizationType.FQ,
         mode=QuantizationMode.MATERIALIZE,
         weight_params=quantization_hparams.WeightQuantizationParams(
-            use_symmetric=is_weight_symmetric
+            use_symmetric=is_weight_symmetric,
+            calculation_dtype=jnp.bfloat16,
         ),
     )
